requirements.txt
peewee
discord
timezonefinder
geopy
requests
pytz
measurement
emoji
pycountry
python-dateutil
pymysql
countryinfo
currencyconverter
praw
covid
weasyprint
pyyaml